requirements.txt
pillow~=10.4.0
opencv-python~=4.10.0.84
numpy~=2.1.2
open3d~=0.19.0